pyproject.toml
[tool.poetry]
name = "agentorg"
version = "0.0.1"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohappyeyeballs = "2.4.3"
aiohttp = "3.10.10"
aiosignal = "1.3.1"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
async-timeout = "4.0.3"
attrs = "24.2.0"
backoff = "2.2.1"
beautifulsoup4 = "4.12.3"
certifi = "2024.8.30"
cffi = "1.17.1"
chardet = "5.2.0"
charset-normalizer = "3.4.0"
click = "8.1.7"
cryptography = "43.0.3"
dataclasses-json = "0.6.7"
distro = "1.9.0"
dnspython = "2.7.0"
email_validator = "2.2.0"
emoji = "2.14.0"
eval_type_backport = "0.2.0"
exceptiongroup = "1.2.2"
fastapi = "0.115.3"
fastapi-cli = "0.0.5"
filetype = "1.2.0"
frozenlist = "1.4.1"
greenlet = "3.1.1"
h11 = "0.14.0"
html5lib = "1.1"
httpcore = "1.0.6"
httptools = "0.6.4"
httpx = "0.27.2"
httpx-sse = "0.4.0"
idna = "3.10"
Jinja2 = "3.1.4"
jiter = "0.6.1"
joblib = "1.4.2"
jsonpatch = "1.33"
jsonpath-python = "1.0.6"
jsonpointer = "3.0.0"
langchain = "0.3.4"
langchain-community = "0.3.3"
langchain-core = "0.3.12"
langchain-openai = "0.2.3"
langchain-text-splitters = "0.3.0"
langdetect = "1.0.9"
langgraph = "0.2.39"
langgraph-checkpoint = "2.0.1"
langgraph-sdk = "0.1.33"
langsmith = "0.1.136"
Levenshtein = "0.26.0"
linkify-it-py = "2.0.3"
lxml = "5.3.0"
markdown-it-py = "3.0.0"
MarkupSafe = "3.0.2"
marshmallow = "3.23.0"
mdit-py-plugins = "0.4.2"
mdurl = "0.1.2"
msgpack = "1.1.0"
multidict = "6.1.0"
mypy-extensions = "1.0.0"
nest-asyncio = "1.6.0"
networkx = "3.4.2"
nltk = "3.9.1"
numpy = "1.26.4"
olefile = "0.47"
openai = "1.55.3"
orjson = "3.10.9"
outcome = "1.3.0.post0"
packaging = "24.1"
phonenumbers = "8.13.48"
platformdirs = "4.3.6"
propcache = "0.2.0"
psutil = "6.1.0"
pycparser = "2.22"
pydantic = "2.9.2"
pydantic-settings = "2.6.0"
pydantic_core = "2.23.4"
Pygments = "2.18.0"
pypdf = "5.1.0"
PySocks = "1.7.1"
python-dateutil = "2.8.2"
python-dotenv = "1.0.1"
python-iso639 = "2024.10.22"
python-Levenshtein = "0.26.0"
python-magic = "0.4.27"
python-multipart = "0.0.12"
python-oxmsg = "0.0.1"
PyYAML = "6.0.2"
RapidFuzz = "3.10.1"
regex = "2024.9.11"
requests = "2.32.3"
requests-toolbelt = "1.0.0"
rich = "13.9.3"
selenium = "4.26.1"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
sortedcontainers = "2.4.0"
soupsieve = "2.6"
SQLAlchemy = "2.0.36"
starlette = "0.41.0"
tenacity = "9.0.0"
textual = "0.85.2"
tiktoken = "0.8.0"
tqdm = "4.66.5"
trio = "0.27.0"
trio-websocket = "0.11.1"
typer = "0.12.5"
typing-inspect = "0.9.0"
typing_extensions = "4.12.2"
uc-micro-py = "1.0.3"
unstructured = "0.16.4"
unstructured-client = "0.27.0"
urllib3 = "2.2.3"
uvicorn = "0.32.0"
uvloop = "0.21.0"
watchfiles = "0.24.0"
webencodings = "0.5.1"
websocket-client = "1.8.0"
websockets = "13.1"
wrapt = "1.16.0"
wsproto = "1.2.0"
yarl = "1.15.5"
faiss-cpu = "1.9.0"
tavily-python = "0.5.0"
webdriver-manager = "4.0.2"
scipy = "1.14.1"
pandas = "2.2.3"
litellm = "1.56.4"
pymilvus = "2.4.7"
mysql-connector-python = "8.3.0"
PyMySQL = "1.1.0"
ShopifyAPI = "12.7.0"
janus = "1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
